Simplify test definition
This commit brings some of the infrastructure I've been using in
NQuery to simplify the test definition. Specifically, instead of
having to guess the spans, this now allows to tag the spans in the
source code using {{ }}.

I've also changed the way diagnostics are specified to a simple
string that has one diagnostic per line, with the ID simply being
separated using a colon. This makes tests look like this:

    var source = @"
        using System;

        namespace ConsoleApp1
        {
            class Program
            {
                static void Main(string[] args)
                {
                    var width = Console.{{WindowWidth}};
                }
            }
        }
    ";

    var expected = @"
        PNS001: Console.WindowWidth isn't supported on Linux, MacOSX
    ";

    AssertMatch(source, expected);

This also simplifies some of the analyzer's template boilerplate to
something a bit more manageable.