Allow analyzer to be customized by project settings
This allows the customer to tweak the analyzer behavior with project
settings. For instance, the project can decide to ignore all issues
for Linux and macOS by adding these settings to the project file:

    <PropertyGroup>
      <PnsIngore>Linux;MacOSX</PnsIngore>
    </PropertyGroup>

This change will also turn off the analyzer unless the project is
targeting .NET Standard or .NET Core.