Add analyzer for deprecated APIs
This adds a generic analyzer that allows deprecating APIs. Adding
deprecated APIs is easy. Simply add a document named DEXXXX.md (where
XXXX is a four digit number) in the docs folder that looks like this:

    <!--
    T:SomeNamespace.SomeType
    -->
    # DEXXXX: Don't use SomeType

    ## Motivation

    `SomeType` is a deprecated APIs. Don't use it or they will be
    dragons.

    ## Recommendation

    The replacement for `SomeType` is the `SomeOtherType`. It's not a
    drop-in replacement, so make sure to consult the [docs](link).

The build will automatically create an index from these documents and
report usages. It will use the doc as the help link.

This fixes #21.