Enable test running on the command line
First, this centralizes the package references that are needed for
xUnit, including the VS and MSBuild test runners.

Secondly, it will run the tests from the command unless tests that is
disabled. Test executing after build will be disabled by AppVeyor (as
AppVeyor already runs them) and when building inside of VS (because
that's already done by Test Explorer).