Add inclusion/exclusion to PNSE set and new DE0008
Add simple inclusion/exclusion capability to ex-gen so the set of PNSE
can be controlled as needed. Opportunistically added also a way to
process previous downloaded versions of .NET Core in order to avoid
downloading on every run - pointed to download .NET Core 2.0.0.

Added deprecation DE0008 that applies to some specific case that can
throw PNSE depending on the value of the parameter on the call. The idea
is to check ApiCompat/Wiki to see if there are more cases like this.