Add inclusion/exclusion to PNSE (#35)
* Add inclusion/exclusion to PNSE

Add simple inclusion/exclusion capability to ex-gen so the set of PNSE
can be controlled as needed. Opportunistically added also a way to
process previous downloaded versions of .NET Core in order to avoid
downloading on every run - pointed to download .NET Core 2.0.0.

Per code review: the two sets of APIs in the added new deprecations
should be in PNSE (the plan being to improve on PNSE).