Add deprecation information for GetTypeInfo() and TypeInfo
We should probably also make the tool smart enough to not display this
information if the System.Type provided by the platform doesn't expose
reflection APIs.

We can do this by either supporting a condition for deprecation that
is based on whether other APIs exist (e.g. Type.GetMembers()) or by
marking this issue as only relevant for specific target frameworks.