Do not include the header in the list of diagnostic IDs
Visual Studio will show all diagnostic IDs in solution explorer so that
the developer can configure them (error, warning, info, off).

We produce the list by parsing the CSV document but we also include the
header row which means we include one diagnostic with an ID of
`DiagnosticIds` which is dumb.