Properties can be assigned via backing field
The code was assuming the presence of a setter if the property was on
the left side of an attribution, but that can be done via a backing
field in some special cases. The analyzer should ignore such cases since
they are not coming from the properties, and respective getters and
setters, that we are interested.

Fixes #96