Change ex-gen to use publish
Right now, we're downloading the SDK from blob storage. That certainly
has its perks, such as not being depending on the .NET Core SDK that
the dev running ex-gen has. However, given that this is basically just
us, this doesn't seem to be too valuable.

However, it also has major downside: we cannot easily add NuGet packages
for indexing, such as Microsoft.Windows.Compatibility.

This changes the flow of how the indexer operates. Instead of
downloading and extracting the SDKs, it generates a dummy project and
publishes it for a specific RID. Adding NuGet packages is easy as it
just requires adding them to the dummy project.