Concrete types of DE0006 should not warn on declaration
The idea here is to prevent false positives like issue #86.

A helper tool is being added to allow to quickly gather the methods that
should be targeted: constructors and static methods of the deprecated
type returning the type. This should be enough to raise warnings for
code creating new instances of the type while allowing usage of APIs
old APIs still returning the types.